+++ test/mld.jl
@@ -10,6 +10,7 @@
 
     @testset "test iterative optimal switching" begin
         @testset "test iterative optimal switching - lindistflow - block" begin
+            @info "    test iterative optimal switching - lindistflow - block"
             args = deepcopy(orig_args)
             args["opt-switch-formulation"] = "lindistflow"
             args["opt-switch-algorithm"] = "iterative"
@@ -21,6 +22,7 @@
         end
 
         @testset "test iterative optimal switching - lindistflow - traditional" begin
+            @info "    test iterative optimal switching - lindistflow - traditional"
             args = deepcopy(orig_args)
             args["opt-switch-formulation"] = "lindistflow"
             args["opt-switch-algorithm"] = "iterative"
@@ -28,10 +30,11 @@
 
             entrypoint(args)
 
-            @test get_timestep_device_actions!(args) == Dict{String, Any}[Dict("Shedded loads" => ["692_3", "675b", "675a", "692_1", "702", "703", "675c"], "Switch configurations" => Dict("801675" => "open", "671692" => "open", "671700" => "open", "703800" => "open", "800801" => "open", "701702" => "open")), Dict("Shedded loads" => ["701", "702", "700", "703"], "Switch configurations" => Dict("801675" => "closed", "671692" => "open", "671700" => "open", "703800" => "open", "800801" => "open", "701702" => "open")), Dict("Shedded loads" => ["702", "703"], "Switch configurations" => Dict("801675" => "closed", "671692" => "open", "671700" => "open", "703800" => "open", "800801" => "open", "701702" => "closed")), Dict("Shedded loads" => String[], "Switch configurations" => Dict("801675" => "closed", "671692" => "closed", "671700" => "open", "703800" => "open", "800801" => "open", "701702" => "closed")), Dict("Shedded loads" => String[], "Switch configurations" => Dict("801675" => "closed", "671692" => "closed", "671700" => "open", "703800" => "open", "800801" => "closed", "701702" => "closed")), Dict("Shedded loads" => ["801"], "Switch configurations" => Dict("801675" => "closed", "671692" => "closed", "671700" => "open", "703800" => "closed", "800801" => "closed", "701702" => "closed")), Dict("Shedded loads" => String[], "Switch configurations" => Dict("801675" => "closed", "671692" => "closed", "671700" => "open", "703800" => "closed", "800801" => "closed", "701702" => "closed")), Dict("Shedded loads" => String[], "Switch configurations" => Dict("801675" => "open", "671692" => "closed", "671700" => "closed", "703800" => "closed", "800801" => "closed", "701702" => "closed"))]
+            @test get_timestep_device_actions!(args) == Dict{String, Any}[Dict("Shedded loads" => ["692_3", "675b", "675a", "692_1", "702", "703", "675c"], "Switch configurations" => Dict("801675" => "open", "671692" => "open", "671700" => "open", "703800" => "open", "800801" => "open", "701702" => "open")), Dict("Shedded loads" => ["701", "702", "700", "703"], "Switch configurations" => Dict("801675" => "closed", "671692" => "open", "671700" => "open", "703800" => "open", "800801" => "open", "701702" => "open")), Dict("Shedded loads" => ["702", "703"], "Switch configurations" => Dict("801675" => "closed", "671692" => "open", "671700" => "open", "703800" => "open", "800801" => "open", "701702" => "closed")), Dict("Shedded loads" => String[], "Switch configurations" => Dict("801675" => "closed", "671692" => "closed", "671700" => "open", "703800" => "open", "800801" => "open", "701702" => "closed")), Dict("Shedded loads" => String[], "Switch configurations" => Dict("801675" => "closed", "671692" => "closed", "671700" => "open", "703800" => "open", "800801" => "closed", "701702" => "closed")), Dict("Shedded loads" => ["801"], "Switch configurations" => Dict("801675" => "closed", "671692" => "closed", "671700" => "open", "703800" => "closed", "800801" => "closed", "701702" => "closed")), Dict("Shedded loads" => String[], "Switch configurations" => Dict("801675" => "closed", "671692" => "closed", "671700" => "open", "703800" => "closed", "800801" => "closed", "701702" => "closed")), Dict("Shedded loads" => String[], "Switch configurations" => Dict("801675" => "closed", "671692" => "closed", "671700" => "open", "703800" => "closed", "800801" => "closed", "701702" => "closed"))]
         end
 
         @testset "test iterative optimal switching - nfa - block" begin
+            @info "    test iterative optimal switching - nfa - block"
             args = deepcopy(orig_args)
             args["opt-switch-formulation"] = "nfa"
             args["opt-switch-algorithm"] = "iterative"
@@ -43,6 +46,7 @@
         end
 
         @testset "test iterative optimal switching - nfa - traditional" begin
+            @info "    test iterative optimal switching - nfa - traditional"
             args = deepcopy(orig_args)
             args["opt-switch-formulation"] = "nfa"
             args["opt-switch-algorithm"] = "iterative"
@@ -56,6 +60,7 @@
 
     @testset "test global optimal switching" begin
         @testset "test global optimal switching - lindistflow - block" begin
+            @info "    test global optimal switching - lindistflow - block"
             args = deepcopy(orig_args)
             args["opt-switch-algorithm"] = "global"
             args["opt-switch-formulation"] = "lindistflow"
@@ -68,6 +73,7 @@
         end
 
         @testset "test global optimal switching - lindistflow - traditional" begin
+            @info "    test global optimal switching - lindistflow - traditional"
             args = deepcopy(orig_args)
             args["opt-switch-algorithm"] = "global"
             args["opt-switch-formulation"] = "lindistflow"
@@ -75,11 +81,12 @@
 
             entrypoint(args)
 
-            @test args["optimal_switching_results"]["5"]["solution"]["load"]["801"]["status"] == PMD.DISABLED
+            @test args["optimal_switching_results"]["6"]["solution"]["load"]["801"]["status"] == PMD.DISABLED
             @test isapprox(args["optimal_switching_results"]["1"]["objective"], 81.07; atol=1)
         end
 
         @testset "test global optimal switching - lindistflow - block - radial-disabled - inverter-disabled" begin
+            @info "    test global optimal switching - lindistflow - block - radial-disabled - inverter-disabled"
             args = deepcopy(orig_args)
             args["opt-switch-algorithm"] = "global"
             args["opt-switch-formulation"] = "lindistflow"
@@ -93,6 +100,7 @@
         end
 
         @testset "test global optimal switching - lindistflow - traditional - radial-disabled - inverter-disabled" begin
+            @info "    test global optimal switching - lindistflow - traditional - radial-disabled - inverter-disabled"
             args = deepcopy(orig_args)
             args["opt-switch-algorithm"] = "global"
             args["opt-switch-formulation"] = "lindistflow"
@@ -106,6 +114,7 @@
         end
 
         @testset "test global optimal switching - nfa - block" begin
+            @info "    test global optimal switching - nfa - block"
             args = deepcopy(orig_args)
             args["opt-switch-algorithm"] = "global"
             args["opt-switch-formulation"] = "nfa"
@@ -117,6 +126,7 @@
         end
 
         @testset "test global optimal switching - nfa - traditional" begin
+            @info "    test global optimal switching - nfa - traditional"
             args = deepcopy(orig_args)
             args["opt-switch-algorithm"] = "global"
             args["opt-switch-formulation"] = "nfa"
@@ -125,6 +135,19 @@
             entrypoint(args)
 
             @test isapprox(args["optimal_switching_results"]["1"]["objective"], 69.66; atol=1)
+        end
+
+        @testset "test robust switching - lindistflow - block" begin
+            @info "    test robust switching - lindistflow - block"
+            args = deepcopy(orig_args)
+            args["opt-switch-formulation"] = "lindistflow"
+            args["opt-switch-algorithm"] = "robust"
+            args["opt-switch-problem"] = "block"
+
+            entrypoint(args)
+
+            @test isapprox(args["optimal_switching_results"]["1"]["objective"], 25.5; atol=1)
+
         end
     end
 end